+++ farmacia/laboratorio.py
@@ -1,0 +1,10 @@
+class Laboratorio:
+
+    def __init__(self, nome: str, endereco: str, telefone: str, cidade: str, estado: str) -> None:
+        self.nome = nome
+        self.endereco = endereco
+        self.telefone = telefone
+        self.cidade = cidade
+        self.estado = estado
+
+
