+++ farmacia/vendas.py
@@ -1,0 +1,11 @@
+from clientes import Cliente
+
+class Vendas:
+
+    def __init__(self, data_hora: str, produtos_vendidos: str, cliente: Cliente, valor_total: float) -> None:
+        self.data_hora = data_hora
+        self.produtos_vendidos = produtos_vendidos
+        self.cliente = Cliente
+        self.valor_total - valor_total
+
+
